+++ api/translateMenu.ts
@@ -1,5 +1,5 @@
 import { GoogleGenerativeAI } from "@google/generative-ai";
-import { extractJsonFromText } from "../apiUtils/data";
+import { extractJsonFromText } from "../apiUtils/parsing";
 import { assignUniqueIds } from "../apiUtils/ids";
 import { GeminiModels, getGeminiApiKey } from "../apiUtils/aiModels";
